requirements-airflow.txt
pyspark==3.5.1
pandas==2.2.2
psycopg2-binary==2.9.9
